split algorithm code into seperate files
