+++ core/src/com/mygdx/game/Node.java
@@ -1,0 +1,28 @@
+package com.mygdx.game;
+
+import com.badlogic.gdx.graphics.Texture;
+import com.badlogic.gdx.graphics.g2d.SpriteBatch;
+
+public class Node {
+    int type;
+    Boolean occupied;
+
+    public Node(int t) {
+        type = t;
+        occupied = false;
+    }
+
+    public void draw(int x, int y, SpriteBatch b) {
+        switch (type) {
+            case 0:
+                b.draw(new Texture(occupied ? "cone occupied.png" : "cone unoccupied.png"), x, y);
+                break;
+            case 1:
+                b.draw(new Texture(occupied ? "cube occupied.png" : "cube unoccupied.png"), x, y);
+                break;
+            case 2:
+                b.draw(new Texture(occupied ? "hybrid occupied.png" : "hybrid unoccupied.png"), x, y);
+                break;
+        }
+    }
+}
